+++ tests/test_key.h
@@ -1,6 +1,8 @@
 #ifndef TEST_KEY_H
 #define TEST_KEY_H
 
-int test_key();
+#include <stdint.h>
+
+uint8_t test_key();
 
 #endif
