+++ tests/test_ntru.h
@@ -1,6 +1,8 @@
 #ifndef TEST_NTRU_H
 #define TEST_NTRU_H
 
-int test_ntru();
+#include <stdint.h>
+
+uint8_t test_ntru();
 
 #endif
